chore: update apr button
