+++ apps/web/src/views/PositionManagers/components/DuoTokenVaultCard.tsx
@@ -154,6 +154,8 @@ export const DuoTokenVaultCard = memo(function DuoTokenVaultCard({
           manager={manager}
           currencyA={currencyA}
           currencyB={currencyB}
+          id={id}
+          totalAssetsInUsd={totalAssetsInUsd}
           earningToken={earningToken}
           price={price}
           vaultName={vaultName}
